[android] network: move implementation to platform folder.
