+++ cmake/treedata/android/subdirs.txt
@@ -3,7 +3,6 @@ xbmc/platform/linux                     platform/linux
 xbmc/input/touch                        input/touch
 xbmc/input/touch/generic                input/touch/generic
 xbmc/network/linux                      network/linux
-xbmc/network/android                    network/android
 xbmc/peripherals/bus/linux              peripherals/bus/linux
 xbmc/peripherals/bus/android            peripherals/bus/android
 xbmc/powermanagement/android            powermanagement/android
@@ -13,4 +12,5 @@ xbmc/windowing/android                  windowing/android
 xbmc/platform/posix                     posix
 xbmc/platform/android/activity          android_activity
 xbmc/platform/android/bionic_supplement android_bionicsupplement
+xbmc/platform/android/network           platform/android/network
 xbmc/platform/android/storage           platform/android/storage
