+++ xbmc/platform/android/network/NetworkAndroid.cpp
