+++ xbmc/platform/android/network/NetworkAndroid.h
