+++ xbmc/platform/android/network/ZeroconfAndroid.cpp
