+++ xbmc/platform/android/network/ZeroconfAndroid.h
