+++ xbmc/platform/android/network/ZeroconfBrowserAndroid.cpp
