+++ xbmc/platform/android/network/ZeroconfBrowserAndroid.h
